Create mixins to require clinician or provider accounts to access a view
Since clinician and provider accounts will generally not have ownership
of the same object types, a problem arises when an account of one type
tries to visit pages designed for the other type. For example, a
provider visiting the Make Commitment page will not have a
ClinicianProfile and will therefore cause a server error. This is by
design because providers should not be making individual commitments.
However, the fact that certain urls can cause a 500 Server Error on
access is bad practice. Instead, we should return a 403 Forbidden to
indicate that the client is at fault for trying to access something it
should not during normal use. The ClinicianLoginRequiredMixin and
ProviderLoginRequiredMixin provide the functionality of the
LoginRequireMixin but also restrict any views they are placed on to
clinician and provider accounts respectively. This commit also applies
them logically to the views to avoid 500 Server Errors. Note that this
solution may be regarded as interim, as while the author is not entirely
 sure that if this is the correct way to implement that behavior in
 Django, the need for a solution is present and the known alternatives
 involve massive code duplication.