Prepare testing layout for more packages
Django custom is to include test files in the the app, as seen in apps
generated with django-admin. While we have extended this by breaking
the tests into multiple files, we generally respect this practice.
Unfortunately, unless we put __init__.py files in the test packages,
pytest will not run two files name test_models.py, even if they reside
in different apps. This commit adds such a file to the commitments test
package. Furthermore, since it creates a package called 'tests', with
the same name as 'tests.py', this commit moves that file into the
package with a rename to keep pytest running.