Make GitHub CI use its default settings for tests
The previous commit makes pytest run with in-memory backends for both
the database and memory. However, this requires the sqlite3 backend, and
sqlite3 has some quirks with filtering strings that will not be present
in our PostgreSQL production database. As it is best to ensure that code
checked into the repo will work in production, the CI instructions
overrule this instruction to pytest in favor of using a database with
the same engine - and therefore behaviors - as production.