Shift behavior of bad delete requests to errorlist
The old behavior of returning a 400 Bad Request on a POST missing the
delete key was convenient for debugging. However, it is generally
incompatible with Django's generic view behavior. Switching to include
errorlist notes in the delete confirmation forms gives as much
information as possible without unusual override of generic views.