Prevent major changes to suggested commitments
When a `Commitment` has a source template, it is from a suggested
commitment. We therefore prevent editing the fields provided by the
template, as well as the course it originated from, as these may change
the nature of the commitment and therefore weaken the quality of
statistics providers can gather.