Remove redundant test
If one looks closely, the test removed was identical to another for the
same view. It was initially included as part of a pair of tests for the
csv when one Course was provided. This test would check the header, the
other would check the first row. However, this strategy was abandoned in
favor of implicitly testing both at once using csv.DictReader. The
header test was incorrectly left in with being edited to actually test
the case where 1 Course is present instead of 0. It can be safely
removed.