Rename main statistics page to be less confusing
Previously, the main provider statistics page was called the statistics
dashboard internally. While this name was not used in the frontend just
yet, it could be confusing to providers if it were, as there would then
be multiple dashboards. Calling it an overview distinguishes it better.
Instances have been renamed as such to keep uniformity with the
inevitable documentation that will refer to it as such.