Relocate suggested commitments table to include
The table for suggested commitments within a course was split, with the
markup for the table itself in the course page template and the markup
for the table data itself in an include. This has two disadvantages.
One, it causes additional clutter in the course page template, which is
already weighty. Two, the division of the table structure makes the
layout opaque to an unfamiliar developer. While there is an argument
that the table data itself should be abstracted out instead of the
cells, that is probably best done by varying includes within the
clinician/provider if statement, and is not mutually exclusive with
relocating the full structure of the table to this one file.