Replace hardcoded http in links
Replaces hardcoded `http` with `{{ request.scheme }}`. This will
maintain http for development but switch to providing https in
production when SSL is enabled. This assumes that the server is either
always-https (it should be in production) or always-http (it should be
in development). This may not work in the (admittedly inadvisable) case
where http and https are mixed.