Add `__getitem__` to CommitmentStatusStatistics
Previously, whenever we wanted to get data from a
CommitmentStatusStatistics object in a template, we had to convert it to
a dictionary first, as Django templates are deliberately light on the
ability to call functions. Implementing `__getitem__` allows the
CommitmentStatusStatistics objects to behave as if they are dictionaries
within the template, while still maintaining non-dictionary behaviors.
This will allow us to add context-dependent statistics (for example,
goodness-of-fit probabilities) or aggregate already-constructed objects.
This will reduce the potential for bloat or code duplication and will
most likely improve performance.