Add a view to clear a Commitment's reminder emails
This view derives from FormView. Using UpdateView with Commitment as the
main object would reduce the amount of overridden code slightly.
However, this would require the accompanying form to inherit from
ModelForm, but not actually modify any of the fields of the object which
it models. This seems like unnecessarily confusing practice. It is
better to suffer a few additional lines of code to avoid such mixed
messaging.