Make commitment status change views UpdateViews
Previously, these views only inherited from View. This worked fine
because they are post-only methods and therefore a manual implementation
of the post method would be about the same in terms of clarity as
adapting UpdateView to our needs. However, now that we need to clear
reminder emails under some status changes, the logic is becoming more
complicated. Refactoring to UpdateView inheritance will allow clearer
code when that change is made.