Make completing & discontinuing commitments cancel reminder emails
If a user completes or discontinues a commitment, we can generally
assume that it has truly been completed or has been abandoned
indefinitely. In that case, it is best to automatically cancel reminder
emails to avoid spamming the clinician. There are a couple ways we could
implement this. One, we could make the CompleteCommitmentForm inherit
from ClearCommitmentReminderEmailsForm. However, this would introduce
rather tight coupling and require an additional field to be submitted in
the post request for the view. Two, we could have CompleteCommitmentForm
call ClearCommitmentReminderEmailsForm when it saves with a commit. This
seems to violate the single responsibility principle pretty badly - it
is a ModelForm that also calls something else as a side effect. Three,
we can have the view make the call to clear emails when it receives a
valid request. This is still a side effect, but since it is intended for
the view as a whole to do this, it is not as bad. In this scenario, the
override of `form_valid` provides clarity about this side effect.
Coupling is also minimized - we always have the option to create another
view that only changes the commitment status without clearing emails,
which is something we would not have if we added this at the form layer.