Move student list to its own context object
While students can be accessed in the template via `course.students`,
the fact that this member is a queryset means it is impossible to
persitently enrich it with anything in a clean way without revision. It
is best to just make students into its own context object, much like
suggested commitments.