Add reminder emails presets to commitment creation
Adds some simple but likely commonly used choices for reminder emails
schedules when first creating a commitment. This also applys when
creating from a suggested commitment. Once a commitment is made,
clinicians will need to edit through 'manage reminder emails'. This will
avoid the possibility of confusion of whether selecting an option during
editing of a commitment will double reminders or not. Additionally,
there is not really a good way to accomodate a custom schedule in the
edit form without risking accidental deletions of an existing schedule.