Fix percentages displaying wrong in popover
Fixes a percent sign with no number showing if there were no commitments
made from a suggested commitment. The section breaking down by status
now only displays if there is at least one commitment made, which is
reasonable because a breakdown is not necessary if the total number made
is zero.